Remove 'open' command
The 'open' command is a rather esoteric command that's most probably
only kept in vim for backwards compatibility. It's very unlikely that it
could be particularly useful for the vast majority of today's users.

Have a look here for more details:
https://vi.stackexchange.com/questions/2275/what-does-open-do-in-vim/2276#2276

Right now it's very prominently placed in the upper left corner of the
cheat sheet. In my opinion, this is confusing, because it leads users to
the wrong assumption that it would be the recommended command for
opening files. Bu in fact 'edit (:e)' should be used for that, which is
already documented in the cheat sheet further down in the "Working with
multiple files" section.

A very practical downside of ':o' compared to ':e' is that the former
one doesn't support tab completion for file paths while the latter one
does.